Inject attempt at expanding heap only on segfault.
This is not viable unless we emit assembler. In order to fix a segfault, we
either need to remap the existing memory, which may be problematic if we run
out of the space reserved for the heap and e.g. into the program text.
Otherwise, we would need to be able to control the value in the register
allocated by the C compiler to the heap pointer and that requires more
control than we can portably have.